to extract the bands in bxsf format
These functions allow to extract the bandstructure in bxsf format and save it in the format used by ELECTRA and EMAF. They also allow to interpolate the bands.

Example of use
bxsf_to_ELECTRA( 'material_fs.bxsf', 'material' , 0.5, 'n', 2) 

In this example, a material_fs.bxsf file is read, converted and stored in a Ek_material.mat file. 0.5 is the lattice parameter in nanometres. 'n' tells the code that SOC has not been used - in the case you use SOC and the system is not magnetic, so that spin up and spin down bands are isdentical, type 'y', differently, you can type whatever you want. An interpolation of a factor of 2 is asked.